requirements.txt
# Core dependencies
pyspark==3.2.4
pydantic>=1.8.0
delta-spark>=1.2.0,<2.0.0
pandas>=1.3.0
numpy>=1.21.0

# Optional dependencies for enhanced functionality
pyarrow>=8.0.0
fastparquet>=0.8.0